feat: gate generic methods behind //go:build go1.27, keep Go 1.18 support
Instead of raising the whole module to go 1.27, the generic methods are now an optional enhancement: go.mod goes back to go 1.18 and every file that needs a method type parameter is guarded by `//go:build go1.27`. That constraint tests the toolchain version, not the go directive, so a Go 1.27 toolchain compiles the generic methods even while go.mod still says 1.18 — verified experimentally before writing any code.

Consequences of that choice, since it drove the layout:

  - The package-level functions can no longer delegate to the methods, because the delegating file might be excluded from the build. Each variant now carries its own implementation. The duplicated bodies are small (a state allocation plus a subscribe), and no shim layer is involved.
  - They are therefore not marked Deprecated either: they are the supported API for Go 1.18 through 1.26, and a fully supported alternative above that.
  - dagcore/dagfunc internals go back to future.Then(f, cb); calling .Then(cb) there would break the Go 1.18 build.
  - dagcore.NodeInstance.Cast and dagfunc.Program.Value/ValueAsync move into their own go1.27-guarded files, together with their tests.
  - dagfunc uses reflect.TypeOf((*T)(nil)).Elem() rather than reflect.TypeFor[T](), because stdversion vet warns about symbols newer than the module's declared language version.
  - CI gains a Go 1.18/1.27 matrix. golangci-lint and Codecov stay on the 1.27 job only, since the linter cannot run on a toolchain older than the one it was built with.

Verified with both toolchains: on Go 1.27, method.go/cast.go/value.go are in the build and all six packages pass with -race plus 0 lint issues; on Go 1.24 those three files are excluded and all six packages still build and pass.
