+++ go.mod
@@ -1,6 +1,6 @@
 module github.com/jizhuozhi/go-future
 
-go 1.27
+go 1.18
 
 require github.com/stretchr/testify v1.9.0
