+++ method_test.go
@@ -1,3 +1,5 @@
+//go:build go1.27
+
 package future
 
 import (
